Read ODF header information

Read ODF header information

read_odf_header(
  file,
  header = odf_parse_header(file, file_encoding = file_encoding),
  file_encoding = "latin1"
)

read_odf_parameter_header(
  file,
  col_types = NULL,
  header = odf_parse_header(file, file_encoding = file_encoding),
  file_encoding = "latin1"
)

read_odf_header_tbl(
  file,
  which,
  col_types = NULL,
  header = odf_parse_header(file, file_encoding = file_encoding),
  file_encoding = "latin1"
)

odf_col_datetime()

odf_header_cols_default(...)

odf_parse_header(
  file,
  header_lines = read_odf_header_lines(file, file_encoding = file_encoding),
  file_encoding = "latin1"
)

read_odf_header_lines(file, n_header = 1000, file_encoding = "latin1")

Arguments

file
A file, URL, or connection. Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed; URLs will be automatically downloaded. See readr::read_lines() for a full description of how this parameter is interpreted.

header
A previously read value obtained from odf_parse_header().

file_encoding
The encoding used to encode the file. The default (latin1) is used to prevent an error involving invalid characters that are common in ODF files.

col_types
A readr::cols() spec or NULL to guess using odf_guess_col_types()

which
The header name (e.g., "PARAMETER_HEADER"). Most ODF files have an ODF_HEADER, a CRUISE_HEADER, EVENT_HEADER, INSTRUMENT_HEADER, HISTORY_HEADER, PARAMETER_HEADER, and RECORD_HEADER.

...
Overrides for the default column types.

header_lines
A previously read value obtained from read_odf_header_lines().

n_header
The starting guess for number of header lines.

Examples

odf_file <- odf_example("CTD_98911_10P_11_DN.ODF")
read_odf_header(odf_file)

#> $ODF_HEADER
#> # A tibble: 1 x 1
#>   FILE_SPECIFICATION                
#>   <chr>                             
#> 1 D1:[MSMITH]CTD_98911_10P_11_DN.ODF
#> 
#> $CRUISE_HEADER
#> # A tibble: 1 x 9
#>   COUNTRY_INSTITUTE_CODE CRUISE_NUMBER ORGANIZATION   CHIEF_SCIENTIST
#>   <chr>                  <chr>         <chr>          <chr>          
#> 1 1810                   98911         OCEAN SCIENCES JH             
#> # … with 5 more variables: START_DATE <dttm>, END_DATE <dttm>, PLATFORM <chr>,
#> #   CRUISE_NAME <chr>, CRUISE_DESCRIPTION <chr>
#> 
#> $EVENT_HEADER
#> # A tibble: 1 x 18
#>   DATA_TYPE EVENT_NUMBER EVENT_QUALIFIER1 EVENT_QUALIFIER2 CREATION_DATE      
#>   <chr>     <chr>        <chr>            <chr>            <dttm>             
#> 1 CTD       10P          11               DN               1998-11-19 16:07:31
#> # … with 13 more variables: ORIG_CREATION_DATE <dttm>, START_DATE_TIME <dttm>,
#> #   END_DATE_TIME <dttm>, INITIAL_LATITUDE <dbl>, INITIAL_LONGITUDE <dbl>,
#> #   END_LATITUDE <dbl>, END_LONGITUDE <dbl>, MIN_DEPTH <dbl>, MAX_DEPTH <dbl>,
#> #   SAMPLING_INTERVAL <dbl>, SOUNDING <dbl>, DEPTH_OFF_BOTTOM <dbl>,
#> #   EVENT_COMMENTS <list>
#> 
#> $INSTRUMENT_HEADER
#> # A tibble: 1 x 4
#>   INST_TYPE MODEL SERIAL_NUMBER DESCRIPTION
#>   <chr>     <chr> <chr>         <chr>      
#> 1 Sea-Bird  NA    253591-026    NA         
#> 
#> $HISTORY_HEADER
#> # A tibble: 3 x 2
#>   CREATION_DATE       PROCESS   
#>   <dttm>              <list>    
#> 1 1998-11-16 10:27:14 <chr [26]>
#> 2 1998-11-19 15:55:53 <chr [8]> 
#> 3 1998-11-19 16:07:31 <chr [15]>
#> 
#> $PARAMETER_HEADER
#> # A tibble: 6 x 14
#>   TYPE  NAME  UNITS    WMO_CODE NULL_VALUE   PRINT_FIELD_WID… PRINT_DECIMAL_PLA…
#>   <chr> <chr> <chr>    <chr>    <chr>                   <dbl>              <dbl>
#> 1 SING  PRES  decibars PRES     -.99000000D…               11                  4
#> 2 SING  TEMP  degrees… TEMP     -.99000000D…               11                  4
#> 3 SING  COND  mmHo     COND     -.99000000D…               11                  4
#> 4 SING  PSAL  NA       PSAL     -.99000000D…               11                  4
#> 5 SING  POTM  degrees… POTM     -.99000000D…                8                  4
#> 6 SING  SIGP  kg/m**3  SIGP     -.99000000D…                8                  4
#> # … with 7 more variables: ANGLE_OF_SECTION <dbl>, MAGNETIC_VARIATION <dbl>,
#> #   DEPTH <chr>, MINIMUM_VALUE <chr>, MAXIMUM_VALUE <chr>, NUMBER_VALID <dbl>,
#> #   NUMBER_NULL <dbl>
#> 
#> $RECORD_HEADER
#> # A tibble: 1 x 3
#>   NUM_HISTORY NUM_CYCLE NUM_PARAM
#>         <dbl>     <dbl>     <dbl>
#> 1           3       562         6
#> 

read_odf_header_tbl(odf_file, "CRUISE_HEADER")

#> # A tibble: 1 x 9
#>   COUNTRY_INSTITUTE_CODE CRUISE_NUMBER ORGANIZATION   CHIEF_SCIENTIST
#>   <chr>                  <chr>         <chr>          <chr>          
#> 1 1810                   98911         OCEAN SCIENCES JH             
#> # … with 5 more variables: START_DATE <dttm>, END_DATE <dttm>, PLATFORM <chr>,
#> #   CRUISE_NAME <chr>, CRUISE_DESCRIPTION <chr>