Read 'ODF' files

ODF (Ocean Data) files are whitespace-delimited files whose column names and are specified in a well-formatted header.

read_odf(
  file,
  col_names = NULL,
  col_types = NULL,
  n_max = -1,
  file_encoding = "latin1"
)

odf_guess_col_names(
  file,
  parameter_header = read_odf_parameter_header(file, file_encoding = file_encoding),
  file_encoding = "latin1"
)

odf_guess_col_types(
  file,
  col_names = odf_guess_col_names(parameter_header = parameter_header),
  parameter_header = read_odf_parameter_header(file, file_encoding = file_encoding),
  file_encoding = "latin1"
)

Arguments

file
A file, URL, or connection. Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed; URLs will be automatically downloaded. See readr::read_lines() for a full description of how this parameter is interpreted.

col_names
A vector of column names or NULL to guess using odf_guess_col_names()

col_types
A readr::cols() spec or NULL to guess using odf_guess_col_types()

n_max
Maximum number of rows to read. read_odf_header_lines().

file_encoding
The encoding used to encode the file. The default (latin1) is used to prevent an error involving invalid characters that are common in ODF files.

parameter_header
A previously read value obtained from read_odf_parameter_header().

Examples

odf_file <- odf_example("CTD_98911_10P_11_DN.ODF")
read_odf(odf_file)

#> # A tibble: 562 x 6
#>     PRES    TEMP  COND  PSAL    POTM  SIGP
#>    <dbl>   <dbl> <dbl> <dbl>   <dbl> <dbl>
#>  1 0.421  0.509   2.67  31.4  0.509   25.2
#>  2 0.8   -0.134   2.68  32.1 -0.134   25.8
#>  3 0.989 -0.229   2.68  32.2 -0.229   25.9
#>  4 1.08  -0.0914  2.68  32.1 -0.0914  25.8
#>  5 1.18   0.317   2.67  31.6  0.317   25.4
#>  6 1.37   0.222   2.67  31.7  0.222   25.4
#>  7 1.37   0.174   2.67  31.8  0.174   25.5
#>  8 1.37  -0.211   2.67  32.1 -0.211   25.8
#>  9 1.46  -0.193   2.66  32.0 -0.193   25.7
#> 10 1.56  -0.622   2.66  32.4 -0.622   26.0
#> # … with 552 more rows

odf_guess_col_names(odf_file)

#> [1] "PRES" "TEMP" "COND" "PSAL" "POTM" "SIGP"

odf_guess_col_types(odf_file)

#> cols(
#>   PRES = col_guess(),
#>   TEMP = col_guess(),
#>   COND = col_guess(),
#>   PSAL = col_guess(),
#>   POTM = col_guess(),
#>   SIGP = col_guess()
#> )