Read ODF header information

Read ODF header information

read_odf_header(
  file,
  header_lines = read_odf_header_lines(file, file_encoding = file_encoding),
  file_encoding = "latin1"
)

read_odf_parameter_header(
  file,
  col_types = readr::cols(),
  header = read_odf_header(file, file_encoding = file_encoding),
  file_encoding = "latin1"
)

read_odf_header_tbl(
  file,
  which,
  col_types = NULL,
  header = read_odf_header(file, file_encoding = file_encoding),
  file_encoding = "latin1"
)

odf_col_datetime()

odf_header_cols_default(...)

read_odf_header_lines(file, n_header = 1000, file_encoding = "latin1")

Arguments

file
A file, URL, or connection. Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed; URLs will be automatically downloaded. See readr::read_lines() for a full description of how this parameter is interpreted.

header_lines
A previously read value obtained from read_odf_header_lines().

file_encoding
The encoding used to encode the file. The default (latin1) is used to prevent an error involving invalid characters that are common in ODF files.

col_types
A readr::cols() spec or NULL to guess using odf_guess_col_types()

header
A previously read value obtained from read_odf_header().

which
The header name (e.g., "PARAMETER_HEADER"). Most ODF files have an ODF_HEADER, a CRUISE_HEADER, EVENT_HEADER, INSTRUMENT_HEADER, HISTORY_HEADER, PARAMETER_HEADER, and RECORD_HEADER.

...
Overrides for the default column types.

n_header
The starting guess for number of header lines.

Examples

odf_file <- odf_example("CTD_98911_10P_11_DN.ODF")
read_odf_parameter_header(odf_file)

#> # A tibble: 6 x 14
#>   TYPE  NAME  UNITS    WMO_CODE NULL_VALUE   PRINT_FIELD_WID… PRINT_DECIMAL_PLA…
#>   <chr> <chr> <chr>    <chr>    <chr>                   <dbl>              <dbl>
#> 1 SING  PRES  decibars PRES     -.99000000D…               11                  4
#> 2 SING  TEMP  degrees… TEMP     -.99000000D…               11                  4
#> 3 SING  COND  mmHo     COND     -.99000000D…               11                  4
#> 4 SING  PSAL  NA       PSAL     -.99000000D…               11                  4
#> 5 SING  POTM  degrees… POTM     -.99000000D…                8                  4
#> 6 SING  SIGP  kg/m**3  SIGP     -.99000000D…                8                  4
#> # … with 7 more variables: ANGLE_OF_SECTION <dbl>, MAGNETIC_VARIATION <dbl>,
#> #   DEPTH <chr>, MINIMUM_VALUE <dbl>, MAXIMUM_VALUE <dbl>, NUMBER_VALID <dbl>,
#> #   NUMBER_NULL <dbl>

read_odf_header_tbl(odf_file, "CRUISE_HEADER")

#> # A tibble: 1 x 9
#>   COUNTRY_INSTITUTE_CODE CRUISE_NUMBER ORGANIZATION   CHIEF_SCIENTIST
#>   <chr>                  <chr>         <chr>          <chr>          
#> 1 1810                   98911         OCEAN SCIENCES JH             
#> # … with 5 more variables: START_DATE <dttm>, END_DATE <dttm>, PLATFORM <chr>,
#> #   CRUISE_NAME <chr>, CRUISE_DESCRIPTION <chr>

str(read_odf_header_lines(odf_file))

#>  chr [1:186] " ODF_HEADER," ...

str(read_odf_header(odf_file))

#> List of 7
#>  $ ODF_HEADER       :List of 1
#>   ..$ :List of 1
#>   .. ..$ FILE_SPECIFICATION: chr "D1:[MSMITH]CTD_98911_10P_11_DN.ODF"
#>  $ CRUISE_HEADER    :List of 1
#>   ..$ :List of 9
#>   .. ..$ COUNTRY_INSTITUTE_CODE: chr "1810"
#>   .. ..$ CRUISE_NUMBER         : chr "98911"
#>   .. ..$ ORGANIZATION          : chr "OCEAN SCIENCES"
#>   .. ..$ CHIEF_SCIENTIST       : chr "JH"
#>   .. ..$ START_DATE            : chr "17-NOV-1858 00:00:00.00"
#>   .. ..$ END_DATE              : chr "17-NOV-1858 00:00:00.00"
#>   .. ..$ PLATFORM              : chr "LOUIS ST. LAURENT"
#>   .. ..$ CRUISE_NAME           : chr "BARROW STRAIT"
#>   .. ..$ CRUISE_DESCRIPTION    : chr ""
#>  $ EVENT_HEADER     :List of 1
#>   ..$ :List of 18
#>   .. ..$ DATA_TYPE         : chr "CTD"
#>   .. ..$ EVENT_NUMBER      : chr "10P"
#>   .. ..$ EVENT_QUALIFIER1  : chr "11"
#>   .. ..$ EVENT_QUALIFIER2  : chr "DN"
#>   .. ..$ CREATION_DATE     : chr "19-NOV-1998 16:07:31.45"
#>   .. ..$ ORIG_CREATION_DATE: chr "16-NOV-1998 10:27:10.02"
#>   .. ..$ START_DATE_TIME   : chr "11-AUG-1998 23:53:19.00"
#>   .. ..$ END_DATE_TIME     : chr "12-AUG-1998 00:02:40.00"
#>   .. ..$ INITIAL_LATITUDE  : chr "74.2734"
#>   .. ..$ INITIAL_LONGITUDE : chr "-90.7461"
#>   .. ..$ END_LATITUDE      : chr "74.2734"
#>   .. ..$ END_LONGITUDE     : chr "-90.7461"
#>   .. ..$ MIN_DEPTH         : chr "0.4"
#>   .. ..$ MAX_DEPTH         : chr "233.6"
#>   .. ..$ SAMPLING_INTERVAL : chr "1.00000000"
#>   .. ..$ SOUNDING          : chr "238.0"
#>   .. ..$ DEPTH_OFF_BOTTOM  : chr "-99.9"
#>   .. ..$ EVENT_COMMENTS    : chr [1:2] "16-NOV-1998 10:27:14 HOSE NOT ON PUMP." "16-NOV-1998 10:27:14 STATION REPEATED AUGUST 12."
#>  $ INSTRUMENT_HEADER:List of 1
#>   ..$ :List of 4
#>   .. ..$ INST_TYPE    : chr "Sea-Bird"
#>   .. ..$ MODEL        : chr ""
#>   .. ..$ SERIAL_NUMBER: chr "253591-026"
#>   .. ..$ DESCRIPTION  : chr ""
#>  $ HISTORY_HEADER   :List of 3
#>   ..$ :List of 2
#>   .. ..$ CREATION_DATE: chr "16-NOV-1998 10:27:14.59"
#>   .. ..$ PROCESS      : chr [1:26] "SELECT_FILE,FILE_SPEC=D1:[MSMITH]BRST9810.CNV" "OPEN_ASCII_FILE" "HEADER_LOGGER" "READ_ASCII_DATA,HEADER_LINES=0," ...
#>   ..$ :List of 2
#>   .. ..$ CREATION_DATE: chr "19-NOV-1998 15:55:53.39"
#>   .. ..$ PROCESS      : chr [1:8] "SELECT_FILE,FI=D1:[MSMITH.DN]CTD_98911*.ODF" "OPEN_ASCII_FILE" "READ_ASCII_HEADERS" "READ_ASCII_DATA,FILE_FORMAT=ODF" ...
#>   ..$ :List of 2
#>   .. ..$ CREATION_DATE: chr "19-NOV-1998 16:07:31.49"
#>   .. ..$ PROCESS      : chr [1:15] "SELECT_FILE,FI=D5:[MSMITH]CTD_98911*.ODF" "OPEN_ASCII_FILE" "READ_ASCII_HEADERS" "READ_ASCII_DATA,FILE_FORMAT=ODF" ...
#>  $ PARAMETER_HEADER :List of 6
#>   ..$ :List of 14
#>   .. ..$ TYPE                : chr "SING"
#>   .. ..$ NAME                : chr "PRES"
#>   .. ..$ UNITS               : chr "decibars"
#>   .. ..$ WMO_CODE            : chr "PRES"
#>   .. ..$ NULL_VALUE          : chr "-.99000000D+02"
#>   .. ..$ PRINT_FIELD_WIDTH   : chr "11"
#>   .. ..$ PRINT_DECIMAL_PLACES: chr "4"
#>   .. ..$ ANGLE_OF_SECTION    : chr "0.0"
#>   .. ..$ MAGNETIC_VARIATION  : chr "0.0"
#>   .. ..$ DEPTH               : chr "0.00000000D+00"
#>   .. ..$ MINIMUM_VALUE       : chr "0.4210"
#>   .. ..$ MAXIMUM_VALUE       : chr "233.6350"
#>   .. ..$ NUMBER_VALID        : chr "562"
#>   .. ..$ NUMBER_NULL         : chr "0"
#>   ..$ :List of 14
#>   .. ..$ TYPE                : chr "SING"
#>   .. ..$ NAME                : chr "TEMP"
#>   .. ..$ UNITS               : chr "degrees C"
#>   .. ..$ WMO_CODE            : chr "TEMP"
#>   .. ..$ NULL_VALUE          : chr "-.99000000D+02"
#>   .. ..$ PRINT_FIELD_WIDTH   : chr "11"
#>   .. ..$ PRINT_DECIMAL_PLACES: chr "4"
#>   .. ..$ ANGLE_OF_SECTION    : chr "0.0"
#>   .. ..$ MAGNETIC_VARIATION  : chr "0.0"
#>   .. ..$ DEPTH               : chr "0.00000000D+00"
#>   .. ..$ MINIMUM_VALUE       : chr "-1.7735"
#>   .. ..$ MAXIMUM_VALUE       : chr "0.5086"
#>   .. ..$ NUMBER_VALID        : chr "562"
#>   .. ..$ NUMBER_NULL         : chr "0"
#>   ..$ :List of 14
#>   .. ..$ TYPE                : chr "SING"
#>   .. ..$ NAME                : chr "COND"
#>   .. ..$ UNITS               : chr "mmHo"
#>   .. ..$ WMO_CODE            : chr "COND"
#>   .. ..$ NULL_VALUE          : chr "-.99000000D+02"
#>   .. ..$ PRINT_FIELD_WIDTH   : chr "11"
#>   .. ..$ PRINT_DECIMAL_PLACES: chr "4"
#>   .. ..$ ANGLE_OF_SECTION    : chr "0.0"
#>   .. ..$ MAGNETIC_VARIATION  : chr "0.0"
#>   .. ..$ DEPTH               : chr "0.00000000D+00"
#>   .. ..$ MINIMUM_VALUE       : chr "2.5819"
#>   .. ..$ MAXIMUM_VALUE       : chr "2.7704"
#>   .. ..$ NUMBER_VALID        : chr "562"
#>   .. ..$ NUMBER_NULL         : chr "0"
#>   ..$ :List of 14
#>   .. ..$ TYPE                : chr "SING"
#>   .. ..$ NAME                : chr "PSAL"
#>   .. ..$ UNITS               : chr ""
#>   .. ..$ WMO_CODE            : chr "PSAL"
#>   .. ..$ NULL_VALUE          : chr "-.99000000D+02"
#>   .. ..$ PRINT_FIELD_WIDTH   : chr "11"
#>   .. ..$ PRINT_DECIMAL_PLACES: chr "4"
#>   .. ..$ ANGLE_OF_SECTION    : chr "0.0"
#>   .. ..$ MAGNETIC_VARIATION  : chr "0.0"
#>   .. ..$ DEPTH               : chr "0.00000000D+00"
#>   .. ..$ MINIMUM_VALUE       : chr "31.4122"
#>   .. ..$ MAXIMUM_VALUE       : chr "33.7702"
#>   .. ..$ NUMBER_VALID        : chr "562"
#>   .. ..$ NUMBER_NULL         : chr "0"
#>   ..$ :List of 14
#>   .. ..$ TYPE                : chr "SING"
#>   .. ..$ NAME                : chr "POTM"
#>   .. ..$ UNITS               : chr "degrees C"
#>   .. ..$ WMO_CODE            : chr "POTM"
#>   .. ..$ NULL_VALUE          : chr "-.99000000D+02"
#>   .. ..$ PRINT_FIELD_WIDTH   : chr "8"
#>   .. ..$ PRINT_DECIMAL_PLACES: chr "4"
#>   .. ..$ ANGLE_OF_SECTION    : chr "0.0"
#>   .. ..$ MAGNETIC_VARIATION  : chr "0.0"
#>   .. ..$ DEPTH               : chr "0.00000000D+00"
#>   .. ..$ MINIMUM_VALUE       : chr "-1.7753"
#>   .. ..$ MAXIMUM_VALUE       : chr "0.5086"
#>   .. ..$ NUMBER_VALID        : chr "562"
#>   .. ..$ NUMBER_NULL         : chr "0"
#>   ..$ :List of 14
#>   .. ..$ TYPE                : chr "SING"
#>   .. ..$ NAME                : chr "SIGP"
#>   .. ..$ UNITS               : chr "kg/m**3"
#>   .. ..$ WMO_CODE            : chr "SIGP"
#>   .. ..$ NULL_VALUE          : chr "-.99000000D+02"
#>   .. ..$ PRINT_FIELD_WIDTH   : chr "8"
#>   .. ..$ PRINT_DECIMAL_PLACES: chr "4"
#>   .. ..$ ANGLE_OF_SECTION    : chr "0.0"
#>   .. ..$ MAGNETIC_VARIATION  : chr "0.0"
#>   .. ..$ DEPTH               : chr "0.00000000D+00"
#>   .. ..$ MINIMUM_VALUE       : chr "25.1871"
#>   .. ..$ MAXIMUM_VALUE       : chr "27.1418"
#>   .. ..$ NUMBER_VALID        : chr "562"
#>   .. ..$ NUMBER_NULL         : chr "0"
#>  $ RECORD_HEADER    :List of 1
#>   ..$ :List of 3
#>   .. ..$ NUM_HISTORY: chr "3"
#>   .. ..$ NUM_CYCLE  : chr "562"
#>   .. ..$ NUM_PARAM  : chr "6"